requirements.txt
streamlit==1.32.0
extra-streamlit-components==0.1.71
groq==0.5.0
qdrant-client==1.8.0
sentence-transformers==2.6.0
python-dotenv==1.0.1
google-api-python-client==2.122.0
pymupdf==1.23.26
python-docx==1.1.0
pandas==2.2.1
openpyxl==3.1.2
langchain-text-splitters==0.0.1
beautifulsoup4==4.12.3
requests==2.31.0
numpy==1.26.4
python==3.10.0